refactor(#297): 스케쥴 관련 유틸 함수 분리
