+++ src/utils/parseUtils.ts
@@ -13,8 +13,6 @@ import {
 import ko from 'date-fns/locale/ko';
 import {
   IClassSchedule,
-  IDayTimeList,
-  IDateTimeList,
   IDaySchedule,
   IMonthlyClassSchedules,
 } from '@/types/class';
@@ -83,76 +81,6 @@ export const getUniqueDates = (dates: Date[]) =>
   dates.filter(
     (date, index, self) => index === self.findIndex((d) => isSameDay(date, d)),
   );
-
-export const calculateFinalDates = (
-  startDate: string,
-  endDate: string,
-  schedules: IDayTimeList[] | IDateTimeList[],
-  holidays: {
-    holiday: string;
-  }[],
-) => {
-  const dayMapping: {
-    [key in '일' | '월' | '화' | '수' | '목' | '금' | '토']: number;
-  } = { 일: 0, 월: 1, 화: 2, 수: 3, 목: 4, 금: 5, 토: 6 };
-
-  let allDates = eachDayOfInterval({
-    start: new Date(startDate),
-    end: new Date(endDate),
-  });
-
-  const holidayDates = holidays.map((holiday) => parseISO(holiday.holiday));
-
-  allDates = allDates.filter(
-    (date) => !holidayDates.some((holidayDate) => isSameDay(date, holidayDate)),
-  );
-
-  if (schedules.length === 0) return [];
-
-  if ('day' in schedules[0]) {
-    const dayTimeSchedules = schedules as IDayTimeList[];
-
-    return dayTimeSchedules.reduce((acc: Date[], schedule) => {
-      const days = schedule.day.map((dayStr) => dayMapping[dayStr]);
-
-      schedule.dateTime.forEach((time) => {
-        const [hourStr, minuteStr] = time.split(':');
-
-        allDates.forEach((date) => {
-          const day = getDay(date);
-
-          if (days.includes(day)) {
-            const hour = parseInt(hourStr, 10);
-            const minute = parseInt(minuteStr, 10);
-            const newDate = set(date, { hours: hour, minutes: minute });
-
-            acc.push(newDate);
-          }
-        });
-      });
-
-      return acc;
-    }, []);
-  } else {
-    const dateTimeSchedules = schedules as IDateTimeList[];
-
-    return dateTimeSchedules.reduce((acc: Date[], schedule) => {
-      schedule.dateTime.forEach((time) => {
-        const [hourStr, minuteStr] = time.split(':');
-        const hour = parseInt(hourStr, 10);
-        const minute = parseInt(minuteStr, 10);
-        const newDate = set(new Date(schedule.date), {
-          hours: hour,
-          minutes: minute,
-        });
-
-        acc.push(newDate);
-      });
-
-      return acc;
-    }, []);
-  }
-};
 
 export const formatPhoneNumber = (phoneNumberString: string) => {
   const cleaned = ('' + phoneNumberString).replace(/\D/g, '');
